Rename package to 'indextools' to avoid collision with existing idxtools package
