+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "idxtools"
+name = "indextools"
 version = "0.1.0"
 description = "A toolkit for accelerating genomics using index files. "
 authors = ["John Didion <jdidion@dnanexus.com>"]
@@ -18,7 +18,7 @@ pytest = "^3.0"
 jinja2 = ["jinja2"]
 
 [tool.poetry.scripts]
-idxtools = "idxtools.console:idxtools"
+indextools = "indextools.console:indextools"
 
 [build-system]
 requires = ["poetry>=0.12"]
